update login-view
the component stores the username and password in its local state. Then, it uses them to fill in two inputs, listens for changes on said inputs, and updates the state accordingly.

Full auth will be implemented later, but for now props.onLoggedIn(username) in the handleSubmit method will allow easy access with any credentials.